+++ wallet/src/main/java/com/mutualfunds/mf/repository/CustomerWalletHistoryRepository.java
@@ -10,13 +10,13 @@ import org.springframework.stereotype.Repository;
 import com.mutualfunds.mf.entity.CustomerWalletHistory;
 
 @Repository
-public interface CustomerWalletHistoryRepository extends JpaRepository<CustomerWalletHistory, Long> {
+public interface CustomerWalletHistoryRepository extends JpaRepository<CustomerWalletHistory, Integer> {
 	
 	@Query(value = "SELECT * FROM customer_wallet_history WHERE CUSTOMER_ID=?", nativeQuery = true)
-	public List<CustomerWalletHistory> getHistories(long customerId);
+	public List<CustomerWalletHistory> getHistories(int customerId);
 	
 	@Modifying
 	@Query(value = "INSERT INTO customer_wallet_history (CUSTOMER_ID, TRANSACTION_DATE, TRANSACTION_TYPE_ID, WALLET_AMOUNT, WALLET_ID) VALUES(?, CURRENT_TIMESTAMP, ?, ?, ?)", nativeQuery = true)
-	public void updateTransactionHistory(long customerId, long transactionTypeId, double walletAmount, long walletId);
+	public void updateTransactionHistory(int customerId, int transactionTypeId, double walletAmount, int walletId);
 	
 }
